typecast to support c++11

+++ VCMTPv3/receiver/vcmtpRecvv3.cpp
@@ -146,11 +146,13 @@ void vcmtpRecvv3::Start()
     mcastHandler();
 }
 
+
 std::string operator+(const std::string& lhs, const struct sockaddr_in& rhs)
 {
     return lhs + inet_ntoa(rhs.sin_addr) + ":" +
-            std::to_string(ntohs(rhs.sin_port));
+            std::to_string(ntohs(static_cast<long long>(rhs.sin_port)));
 }
+
 
 /**
  * Join multicast group specified by mcastAddr:mcastPort.
